// src/dmmf/types.ts
export namespace DMMF {
  export interface Document {
    datamodel: Datamodel;
    schema: Schema;
    modelMappings: ModelMapping[];
  }
  export interface Enum {
    name: string;
    dbName?: string | null;
    typeName: string;
    docs: string | undefined;
    valuesMap: Array<{ name: string; value: string }>;
  }
  export interface Datamodel {
    models: Model[];
    enums: Enum[];
    types: Model[];
  }
  export interface UniqueIndex {
    name: string;
    fields: string[];
  }
  export interface PrimaryKey {
    name: string | null;
    fields: string[];
  }
  export interface Model {
    name: string;
    dbName: string | null;
    fields: ModelField[];
    uniqueFields: string[][];
    uniqueIndexes: UniqueIndex[];
    primaryKey: PrimaryKey | null;
    typeName: string;
    docs: string | undefined;
  }
  export type FieldKind = 'scalar' | 'object' | 'enum' | 'unsupported';
  export type FieldNamespace = 'model' | 'prisma';
  export type FieldLocation =
    | 'scalar'
    | 'inputObjectTypes'
    | 'outputObjectTypes'
    | 'enumTypes';
  // Field
  export interface ModelField {
    // kind: FieldKind;
    name: string;
    isRequired: boolean;
    isList: boolean;
    isUnique: boolean;
    isId: boolean;
    isReadOnly: boolean;
    isGenerated?: boolean;
    isUpdatedAt?: boolean;
    // type: string;
    dbNames?: string[] | null;
    hasDefaultValue: boolean;
    default?: FieldDefault | FieldDefaultScalar | FieldDefaultScalar[];
    relationToFields?: any[];
    relationOnDelete?: string;
    relationName?: string;
    type: string;
    location: FieldLocation;
    typeFieldAlias?: string;
    typeGraphQLType: string;
    fieldTSType: string;
    docs: string | undefined;
    isOmitted: { input: boolean | InputOmitSetting[]; output: boolean };
  }
// ...
}